+++ includes/class-main.php
@@ -81,6 +81,13 @@ class Simple_Events_Calendar {
     public $docs;
 
     /**
+     * Pro upsell UI
+     *
+     * @var Simple_Events_Pro_Upsell
+     */
+    public $pro_upsell;
+
+    /**
      * "Add to Calendar" .ics generator
      *
      * @var Simple_Events_ICS
@@ -199,6 +206,7 @@ class Simple_Events_Calendar {
         require_once PLUGIN_DIR . '/includes/class-meta-box.php';
         require_once PLUGIN_DIR . '/includes/class-settings.php';
         require_once PLUGIN_DIR . '/includes/class-docs.php';
+        require_once PLUGIN_DIR . '/includes/class-pro-upsell.php';
         require_once PLUGIN_DIR . '/includes/class-ics.php';
         require_once PLUGIN_DIR . '/includes/class-templates.php';
         require_once PLUGIN_DIR . '/includes/class-recurrence.php';
@@ -213,6 +221,7 @@ class Simple_Events_Calendar {
         $this->meta_box = new Simple_Events_Meta_Box();
         $this->settings = new Simple_Events_Settings();
         $this->docs = new Simple_Events_Docs();
+        $this->pro_upsell = new Simple_Events_Pro_Upsell();
         $this->ics = new Simple_Events_ICS();
         $this->templates = new Simple_Events_Templates();
         $this->recurrence = new Simple_Events_Recurrence();